requirements.txt
Django>=4.2
torch>=2.0
torchvision>=0.15
numpy
opencv-python
Pillow
